+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
+### Added
+- Add party listing API and party labeling support.
 
 ### Changed
 - New Satori cache flag configuration parameter to require disabling by default.